requirements-dev.txt
# Dev/test deps (keep light; match Makefile install)
pytest
pyyaml
jsonschema>=4.22,<5.0
requests>=2.31,<3.0
numpy>=1.26.4,<2.0
pandas>=2.2.2,<2.3
matplotlib>=3.8,<3.9
